refactor UI into tabs
