+++ node_launcher/gui/launch_widget.py
@@ -6,14 +6,15 @@ from PySide2.QtWidgets import QGridLayout, QMessageBox
 
 from node_launcher.constants import NODE_LAUNCHER_RELEASE, UPGRADE, \
     OPERATING_SYSTEM, LINUX
+from node_launcher.gui.components.tabs import Tabs
 from node_launcher.gui.data_directory import DataDirectoryBox
-from node_launcher.gui.network_group_box import NetworkGroupBox
+from node_launcher.gui.network_buttons import NetworkWidget
 from node_launcher.services.launcher_software import LauncherSoftware
 
 
 class LaunchWidget(QtWidgets.QWidget):
-    mainnet_group_box: NetworkGroupBox
-    testnet_group_box: NetworkGroupBox
+    mainnet_group_box: NetworkWidget
+    testnet_group_box: NetworkWidget
     network_grid: QGridLayout
 
     def __init__(self):
@@ -21,20 +22,20 @@ class LaunchWidget(QtWidgets.QWidget):
         self.message_box = QMessageBox(self)
         self.message_box.setTextFormat(Qt.RichText)
 
-        self.testnet_group_box = NetworkGroupBox(network='testnet',
-                                                 parent=self)
-        self.mainnet_group_box = NetworkGroupBox(network='mainnet',
-                                                 parent=self)
-        self.data_directory_group_box = DataDirectoryBox(self.mainnet_group_box.node_set)
+        self.testnet_group_box = NetworkWidget(network='testnet',
+                                               parent=self)
+        self.mainnet_group_box = NetworkWidget(network='mainnet',
+                                               parent=self)
+        self.data_directory_group_box = DataDirectoryBox(
+            self.mainnet_group_box.node_set)
 
-        self.network_grid = QtWidgets.QGridLayout()
-        self.network_grid.addWidget(self.testnet_group_box, 1, 1)
-        self.network_grid.addWidget(self.mainnet_group_box, 1, 2)
+        self.network_grid = Tabs(mainnet=self.mainnet_group_box,
+                                 testnet=self.testnet_group_box)
 
         self.grid = QtWidgets.QVBoxLayout()
         self.grid.addStretch()
         self.grid.addWidget(self.data_directory_group_box)
-        self.grid.addLayout(self.network_grid)
+        self.grid.addWidget(self.network_grid)
         self.grid.setAlignment(self.data_directory_group_box, Qt.AlignHCenter)
         self.setLayout(self.grid)
