+++ node_launcher/gui/network_buttons/__init__.py
@@ -1,0 +1,1 @@
+from .network_widget import NetworkWidget
