+++ node_launcher/gui/network_buttons/cli_layout.py
@@ -1,6 +1,6 @@
 from node_launcher.gui.components.copy_button import CopyButton
 from node_launcher.gui.components.layouts import QGridLayout
-from node_launcher.gui.network_group_box.section_name import SectionName
+from node_launcher.gui.network_buttons.section_name import SectionName
 from node_launcher.node_set import NodeSet
