+++ node_launcher/gui/network_buttons/lnd_wallet_layout.py
@@ -8,7 +8,7 @@ from grpc._channel import _Rendezvous
 from node_launcher.constants import keyring
 from node_launcher.gui.components.layouts import QGridLayout
 from node_launcher.gui.horizontal_line import HorizontalLine
-from node_launcher.gui.network_group_box.section_name import SectionName
+from node_launcher.gui.network_buttons.section_name import SectionName
 from node_launcher.gui.seed_dialog import SeedDialog
 from node_launcher.node_set import NodeSet
