--- node_launcher/gui/network_group_box/__init__.py
@@ -1,1 +1,0 @@
-from .network_group_box import NetworkGroupBox
